// gravitee-gamma/gravitee-gamma-module-apim/src/main/ui/features/apis/pages/detail/deployment/types.ts
/*
 * Copyright © 2015 The Gravitee team (http://gravitee.io)
 * 
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 * 
 *     http://www.apache.org/licenses/LICENSE-2.0
 * 
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
export type DiffMode = 'line-by-line' | 'side-by-side';

export type UnifiedLine = { type: '+' | '-' | ' '; text: string };

export type SideLine = {
    left: string | null;
    right: string | null;
    kind: 'equal' | 'changed' | 'added' | 'removed';
};